Save the list of build options in misc/options.txt
This commit adds a sofa_option() CMake function to declare options instead
of the standard option(). It stores a list of the options in the cache, as
well as their default values, and the list of options that changed is saved
to 'misc/options.txt' in the build directory at the end of a cmake run.

This should be a convenient way to share your configuration with people you
work with, and a good way to complete your bug reports :-)