cmake: Redo the remove_duplicate from 10889; doesn't cause a weird bug.
Still, this reveals a bug: the first time I run cmake after enabling
SOFA-MISC_TESTS, the include directories for gtest are somehow
forgotten, and the compilation fails. The problem disappears if I run
cmake a second time. It was hidden until now because cmake choked on an
empty list when trying to REMOVE_DUPLICATE, and thus was always run a
second time.